Upstream version 3.0.3

# gpg: Signature made su 21. tammikuuta 2018 17.48.55 EET
# gpg:                using RSA key BED8449FCEE8DA88
# gpg: Good signature from "Otto Kekäläinen <otto@seravo.fi>" [full]
# gpg:                 aka "Otto Kekäläinen <otto@kekalainen.net>" [full]
# gpg:                 aka "Otto Kekäläinen <otto@fsfe.org>" [full]
# gpg:                 aka "Otto Kekäläinen <otto@linux.com>" [full]
# gpg:                 aka "Otto Kekäläinen <otto.kekalainen@seravo.fi>" [full]
# gpg:                 aka "Otto Kekäläinen <otto@mariadb.org>" [undefined]
# gpg:                 aka "Otto Kekäläinen <otto@debian.org>" [undefined]
# gpg:                 aka "Otto Kekäläinen <otto@sange.fi>" [full]

+++ unittest/libmariadb/bulk1.c
@@ -69,7 +69,7 @@ static int bulk1(MYSQL *mysql)
   rc= mysql_query(mysql, "CREATE TABLE bulk1 (a int , b VARCHAR(255))");
   check_mysql_rc(rc, mysql);
 
-  rc= mysql_stmt_prepare(stmt, stmt_str, strlen(stmt_str));
+  rc= mysql_stmt_prepare(stmt, SL(stmt_str));
   check_stmt_rc(rc, stmt);
 
   /* allocate memory */
@@ -155,7 +155,7 @@ static int bulk2(MYSQL *mysql)
   rc= mysql_query(mysql, "CREATE TABLE bulk2 (a int default 4, b int default 2)");
   check_mysql_rc(rc, mysql);
 
-  rc= mysql_stmt_prepare(stmt, "INSERT INTO bulk2 VALUES (?,1)", -1);
+  rc= mysql_stmt_prepare(stmt, SL("INSERT INTO bulk2 VALUES (?,1)"));
   check_stmt_rc(rc, stmt);
 
   memset(bind, 0, 2 * sizeof(MYSQL_BIND));
@@ -211,7 +211,7 @@ static int bulk3(MYSQL *mysql)
   rc= mysql_query(mysql, "CREATE TABLE bulk3 (name varchar(20), row int)");
   check_mysql_rc(rc,mysql);
 
-  rc= mysql_stmt_prepare(stmt, "INSERT INTO bulk3 VALUES (?,?)", -1);
+  rc= mysql_stmt_prepare(stmt, SL("INSERT INTO bulk3 VALUES (?,?)"));
   check_stmt_rc(rc, stmt);
 
   memset(bind, 0, sizeof(MYSQL_BIND)*2);
@@ -265,7 +265,7 @@ static int bulk4(MYSQL *mysql)
   rc= mysql_query(mysql, "CREATE TABLE bulk4 (name varchar(20), row int not null default 3)");
   check_mysql_rc(rc,mysql);
 
-  rc= mysql_stmt_prepare(stmt, "INSERT INTO bulk4 VALUES (?,?)", -1);
+  rc= mysql_stmt_prepare(stmt, SL("INSERT INTO bulk4 VALUES (?,?)"));
   check_stmt_rc(rc, stmt);
 
   memset(bind, 0, sizeof(MYSQL_BIND)*2);
@@ -379,7 +379,7 @@ static int bulk5(MYSQL *mysql)
 
   memset(bind, 0, sizeof(MYSQL_BIND) * 3);
 
-  rc= mysql_stmt_prepare(stmt, "UPDATE bulk5 SET a=? WHERE a=?", -1);
+  rc= mysql_stmt_prepare(stmt, SL("UPDATE bulk5 SET a=? WHERE a=?"));
   check_stmt_rc(rc, stmt);
 
   bind[0].buffer_type= MYSQL_TYPE_LONG;
@@ -443,7 +443,7 @@ static int bulk6(MYSQL *mysql)
   memset(bind, 0, sizeof(MYSQL_BIND) * 3);
 
   /* 1st case: UPDATE */
-  rc= mysql_stmt_prepare(stmt, "UPDATE bulk6 SET a=?, b=? WHERE a=?", -1);
+  rc= mysql_stmt_prepare(stmt, SL("UPDATE bulk6 SET a=?, b=? WHERE a=?"));
   check_stmt_rc(rc, stmt);
 
   bind[0].buffer_type= MYSQL_TYPE_LONG;
@@ -479,7 +479,7 @@ static int bulk6(MYSQL *mysql)
   check_mysql_rc(rc, mysql);
 
   stmt= mysql_stmt_init(mysql);
-  rc= mysql_stmt_prepare(stmt, "INSERT INTO bulk6 VALUES (?,?)", -1);
+  rc= mysql_stmt_prepare(stmt, SL("INSERT INTO bulk6 VALUES (?,?)"));
   check_stmt_rc(rc, stmt);
 
   rc= mysql_stmt_attr_set(stmt, STMT_ATTR_ARRAY_SIZE, &array_size);
@@ -539,6 +539,8 @@ static int test_conc243(MYSQL *mysql)
   size_t row_size= sizeof(struct st_data);
   int rc;
 
+  if (!bulk_enabled)
+    return SKIP;
   rc= mysql_query(mysql, "DROP TABLE IF EXISTS bulk_example2");
   check_mysql_rc(rc, mysql);
 
@@ -547,7 +549,7 @@ static int test_conc243(MYSQL *mysql)
   check_mysql_rc(rc, mysql);
 
   stmt= mysql_stmt_init(mysql);
-  rc= mysql_stmt_prepare(stmt, "INSERT INTO bulk_example2 VALUES (?,?,?)", -1);
+  rc= mysql_stmt_prepare(stmt, SL("INSERT INTO bulk_example2 VALUES (?,?,?)"));
   check_stmt_rc(rc, stmt);
 
   memset(bind, 0, sizeof(MYSQL_BIND) * 3);
@@ -588,7 +590,7 @@ static int test_conc243(MYSQL *mysql)
   if (strcmp(row[0], "Monty") || strcmp(row[1], "Widenius"))
   {
     mysql_free_result(result);
-    diag("Wrong walues");
+    diag("Wrong values");
     return FAIL;
   }
   mysql_free_result(result);
@@ -609,7 +611,7 @@ static int bulk7(MYSQL *mysql)
   rc= mysql_query(mysql, "INSERT INTO t1 VALUES (1)");
   check_mysql_rc(rc, mysql);
 
-  rc= mysql_stmt_prepare(stmt, "UPDATE t1 SET a=a+1", -1);
+  rc= mysql_stmt_prepare(stmt, SL("UPDATE t1 SET a=a+1"));
   check_stmt_rc(rc, stmt);
 
   rc= mysql_stmt_attr_set(stmt, STMT_ATTR_ARRAY_SIZE, &array_size);
@@ -643,7 +645,7 @@ static int test_char_conv1(MYSQL *mysql)
   rc= mysql_query(mysql, "CREATE TABLE char_conv (a varchar(20)) CHARSET=latin1");
   check_mysql_rc(rc, mysql);
 
-  rc= mysql_stmt_prepare(stmt, "INSERT INTO char_conv VALUES (?)", -1);
+  rc= mysql_stmt_prepare(stmt, SL("INSERT INTO char_conv VALUES (?)"));
   check_stmt_rc(rc, stmt);
 
   memset(&bind_in, 0, sizeof(MYSQL_BIND));
@@ -661,7 +663,7 @@ static int test_char_conv1(MYSQL *mysql)
 
   stmt= mysql_stmt_init(mysql);
 
-  rc= mysql_stmt_prepare(stmt, "SELECT a from char_conv", -1);
+  rc= mysql_stmt_prepare(stmt, SL("SELECT a from char_conv"));
   check_stmt_rc(rc, stmt);
 
   memset(&bind_out, 0, sizeof(MYSQL_BIND));
@@ -703,6 +705,9 @@ static int test_char_conv2(MYSQL *mysql)
   char *buffer[1];
   char outbuffer[100];
   
+  if (!bulk_enabled)
+    return SKIP;
+
   buffer[0]= calloc(1, 7);
   strcpy (buffer[0], "\xC3\x82\xC3\x83\xC3\x84\x00");
 
@@ -713,7 +718,7 @@ static int test_char_conv2(MYSQL *mysql)
   rc= mysql_query(mysql, "CREATE TABLE char_conv (a varchar(20)) CHARSET=latin1");
   check_mysql_rc(rc, mysql);
 
-  rc= mysql_stmt_prepare(stmt, "INSERT INTO char_conv VALUES (?)", -1);
+  rc= mysql_stmt_prepare(stmt, SL("INSERT INTO char_conv VALUES (?)"));
   check_stmt_rc(rc, stmt);
 
   rc= mysql_stmt_attr_set(stmt, STMT_ATTR_ARRAY_SIZE, &array_size);
@@ -734,7 +739,7 @@ static int test_char_conv2(MYSQL *mysql)
 
   stmt= mysql_stmt_init(mysql);
 
-  rc= mysql_stmt_prepare(stmt, "SELECT a from char_conv", -1);
+  rc= mysql_stmt_prepare(stmt, SL("SELECT a from char_conv"));
   check_stmt_rc(rc, stmt);
 
   memset(&bind_out, 0, sizeof(MYSQL_BIND));
@@ -767,6 +772,95 @@ static int test_char_conv2(MYSQL *mysql)
   return OK;
 }
 
+
+static int bulk_skip_row(MYSQL *mysql)
+{
+  MYSQL_STMT *stmt;
+  MYSQL_BIND bind[3];
+  MYSQL_RES  *result;
+  MYSQL_ROW  row;
+
+  struct st_data {
+    unsigned long id;
+    char id_ind;
+    char forename[30];
+    char forename_ind;
+    char surname[30];
+    char surname_ind;
+  };
+
+  struct st_data data[]={
+    { 0, STMT_INDICATOR_NULL, "Monty", STMT_INDICATOR_NTS, "Widenius", STMT_INDICATOR_IGNORE_ROW },
+    { 0, STMT_INDICATOR_IGNORE_ROW, "David", STMT_INDICATOR_NTS, "Axmark", STMT_INDICATOR_NTS },
+    { 0, STMT_INDICATOR_NULL, "default", STMT_INDICATOR_DEFAULT, "N.N.", STMT_INDICATOR_NTS },
+  };
+
+  unsigned int array_size= 3;
+  size_t row_size= sizeof(struct st_data);
+  int rc;
+
+  if (!bulk_enabled)
+    return SKIP;
+
+  rc= mysql_query(mysql, "DROP TABLE IF EXISTS bulk_example2");
+  check_mysql_rc(rc, mysql);
+
+  rc= mysql_query(mysql, "CREATE TABLE bulk_example2 (id INT NOT NULL AUTO_INCREMENT PRIMARY KEY,"\
+    "forename CHAR(30) NOT NULL DEFAULT 'unknown', surname CHAR(30))");
+  check_mysql_rc(rc, mysql);
+
+  stmt= mysql_stmt_init(mysql);
+  rc= mysql_stmt_prepare(stmt, SL("INSERT INTO bulk_example2 VALUES (?,?,?)"));
+  check_stmt_rc(rc, stmt);
+
+  memset(bind, 0, sizeof(MYSQL_BIND) * 3);
+
+  /* We autogenerate id's, so all indicators are STMT_INDICATOR_NULL */
+  bind[0].u.indicator= &data[0].id_ind;
+  bind[0].buffer_type= MYSQL_TYPE_LONG;
+
+  bind[1].buffer= &data[0].forename;
+  bind[1].buffer_type= MYSQL_TYPE_STRING;
+  bind[1].u.indicator= &data[0].forename_ind;
+
+  bind[2].buffer_type= MYSQL_TYPE_STRING;
+  bind[2].buffer= &data[0].surname;
+  bind[2].u.indicator= &data[0].surname_ind;
+
+  /* set array size */
+  mysql_stmt_attr_set(stmt, STMT_ATTR_ARRAY_SIZE, &array_size);
+
+  /* set row size */
+  mysql_stmt_attr_set(stmt, STMT_ATTR_ROW_SIZE, &row_size);
+
+  /* bind parameter */
+  mysql_stmt_bind_param(stmt, bind);
+
+  /* execute */
+  rc= mysql_stmt_execute(stmt);
+  check_stmt_rc(rc, stmt);
+
+  mysql_stmt_close(stmt);
+
+  rc= mysql_query(mysql, "SELECT forename, surname FROM bulk_example2");
+  check_mysql_rc(rc, mysql);
+
+  result= mysql_store_result(mysql);
+  FAIL_IF(!result || mysql_num_rows(result) != 1, "Invalid resultset");
+  
+  row = mysql_fetch_row(result);
+  if (strcmp(row[0], "unknown") || strcmp(row[1], "N.N."))
+  {
+    mysql_free_result(result);
+    diag("Wrong values");
+    return FAIL;
+  }
+  mysql_free_result(result);
+  rc= mysql_query(mysql, "DROP TABLE bulk_example2");
+  check_mysql_rc(rc, mysql);
+  return OK;
+}
+
 struct my_tests_st my_tests[] = {
   {"check_bulk", check_bulk, TEST_CONNECTION_DEFAULT, 0,  NULL,  NULL},
   {"test_char_conv1", test_char_conv1, TEST_CONNECTION_NEW, 0,  NULL,  NULL},
@@ -780,6 +874,7 @@ struct my_tests_st my_tests[] = {
   {"bulk3", bulk3, TEST_CONNECTION_DEFAULT, 0,  NULL,  NULL},
   {"bulk4", bulk4, TEST_CONNECTION_DEFAULT, 0,  NULL,  NULL},
   {"bulk_null", bulk_null, TEST_CONNECTION_DEFAULT, 0,  NULL,  NULL},
+  {"bulk_skip_row", bulk_skip_row, TEST_CONNECTION_DEFAULT, 0,  NULL,  NULL},
   {NULL, NULL, 0, 0, NULL, NULL}
 };
